# openapi.yaml
openapi: 3.1.0
info:
    title: Bonyan-API
    version: 3.0.0
    summary: Multi-source fallback REST API for Islamic digital content.
    license:
        name: MIT
        identifier: MIT
    contact:
        name: BonyanOSS
        url: https://github.com/BonyanOSS/Bonyan-API
    description:
        Ordered, validated source adapters. Required fields are never null or placeholders. Missing optional metadata is omitted, never inferred.
        Reciter IDs retain MP3Quran identity; equivalent audio is matched by narrator, narration and style. See SOURCES.md for source scope and
        failover limits.
servers:
    - url: http://localhost:3000
      description: Local development
...
paths:
...
    /prayer/times:
        get:
            tags:
                - Prayer
            summary: Get prayer times
            parameters:
                - name: date
                  in: query
                  schema:
                      type: string
                      pattern: ^\d{2}-\d{2}-\d{4}$
                - name: latitude
                  in: query
                  schema:
                      type: number
                      minimum: -90
                      maximum: 90
                - name: longitude
                  in: query
                  schema:
                      type: number
                      minimum: -180
                      maximum: 180
                - name: city
                  in: query
                  schema:
                      type: string
                - name: country
                  in: query
                  schema:
                      type: string
                - name: method
                  in: query
                  schema:
                      type: integer
                      enum:
                          - 1
                          - 2
                          - 3
                          - 4
                          - 5
                          - 9
                          - 10
                          - 11
                      default: 4
                - name: timezone
                  in: query
                  description: IANA timezone applied to all sources. Coordinate requests support local fallback; city-only requests need the remote geocoder.
                  schema:
                      type: string
                      default: UTC
                      example: Asia/Riyadh
            responses:
                '200':
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PrayerTimesResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '429':
                    $ref: '#/components/responses/RateLimited'
                '500':
                    $ref: '#/components/responses/InternalError'
                '503':
                    $ref: '#/components/responses/Unavailable'
            operationId: getPrayerTimes
...
components:
...
    responses:
        BadRequest:
            description: Invalid request
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
...
        Unavailable:
            description: All eligible sources failed
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        RateLimited:
            description: Rate limit exceeded
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        InternalError:
            description: Internal server error
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
    schemas:
...
        Coordinates:
            type: object
            additionalProperties: false
            required:
                - latitude
                - longitude
            properties:
                latitude:
                    type: number
                    minimum: -90
                    maximum: 90
                longitude:
                    type: number
                    minimum: -180
                    maximum: 180
        PrayerTimings:
            type: object
            additionalProperties: false
            required:
                - date
                - timezone
                - timings
                - method
                - coordinates
                - apiName
            properties:
                date:
                    type: string
                    pattern: ^\d{2}-\d{2}-\d{4}$
                hijri:
                    type: string
                    minLength: 1
                timezone:
                    type: string
                    minLength: 1
                timings:
                    type: object
                    additionalProperties: false
                    required:
                        - Fajr
                        - Sunrise
                        - Dhuhr
                        - Asr
                        - Sunset
                        - Maghrib
                        - Isha
                    properties:
                        Fajr:
                            type: string
                            pattern: ^(?:[01]\d|2[0-3]):[0-5]\d$
                        Sunrise:
                            type: string
                            pattern: ^(?:[01]\d|2[0-3]):[0-5]\d$
                        Dhuhr:
                            type: string
                            pattern: ^(?:[01]\d|2[0-3]):[0-5]\d$
                        Asr:
                            type: string
                            pattern: ^(?:[01]\d|2[0-3]):[0-5]\d$
                        Sunset:
                            type: string
                            pattern: ^(?:[01]\d|2[0-3]):[0-5]\d$
                        Maghrib:
                            type: string
                            pattern: ^(?:[01]\d|2[0-3]):[0-5]\d$
                        Isha:
                            type: string
                            pattern: ^(?:[01]\d|2[0-3]):[0-5]\d$
                method:
                    type: string
                    minLength: 1
                coordinates:
                    $ref: '#/components/schemas/Coordinates'
                apiName:
                    type: string
                    enum:
                        - aladhan.com
                        - local
...
        ErrorResponse:
            type: object
            additionalProperties: false
            required:
                - success
                - message
                - error
            properties:
                success:
                    type: boolean
                    const: false
                message:
                    type: string
                    minLength: 1
                error:
                    type: object
                    additionalProperties: false
                    required:
                        - code
                        - message
                        - requestId
                    properties:
                        code:
                            type: string
                            enum:
                                - BAD_REQUEST
                                - NOT_FOUND
                                - RATE_LIMITED
                                - ALL_SOURCES_FAILED
                                - INTERNAL_SERVER_ERROR
                        message:
                            type: string
                            minLength: 1
                        requestId:
                            type: string
                            minLength: 1
...
        PrayerTimesResponse:
            type: object
            additionalProperties: false
            required:
                - success
                - data
            properties:
                success:
                    type: boolean
                    const: true
                data:
                    $ref: '#/components/schemas/PrayerTimings'
...